Improve React components structure, fix checkRateLimit method response, update some modules
