+++ package.json
@@ -8,7 +8,7 @@
     "url": "git://github.com/Soundnode/soundnode-app.git"
   },
   "scripts": {
-    "prestart": "npm run webpack:prod && npm run sass:prod",
+    "prestart": "npm run watch",
     "start": "./node_modules/.bin/electron .",
     "build": "npm run clean && npm run webpack:prod && npm run sass:prod && npm run package:all",
     "watch": "npm run webpack:dev & npm run sass:dev",
@@ -25,19 +25,18 @@
   "author": "Michael Lancaster",
   "license": "GPL-3.0",
   "devDependencies": {
-    "babel-core": "^6.23.1",
-    "babel-loader": "^6.3.1",
+    "babel-core": "6.24.1",
+    "babel-loader": "^7.0.0",
     "babel-preset-es2015": "^6.22.0",
     "babel-preset-react": "^6.23.0",
     "babel-preset-stage-0": "^6.22.0",
     "babel-register": "^6.23.0",
     "electron-packager": "^8.5.1",
-    "electron": "^1.4.13",
-    "eslint": "^3.15.0",
-    "eslint-plugin-react": "^6.9.0",
-    "install": "^0.8.7",
+    "electron": "^1.6.11",
+    "eslint": "^4.0.0",
+    "eslint-plugin-react": "^7.1.0",
+    "install": "^0.10.1",
     "node-sass": "^4.5.0",
-    "npm": "^4.2.0",
     "rimraf": "^2.5.4",
     "webpack": "^1.12.2"
   },
@@ -49,15 +48,15 @@
     "angular-ui-router": "^0.4.2",
     "electron-window-state": "^4.0.2",
     "font-awesome": "^4.7.0",
-    "fs-extra": "^2.0.0",
+    "fs-extra": "^3.0.1",
     "jquery": "^3.1.1",
     "lodash": "^4.17.4",
     "moment": "^2.17.1",
     "ng-dialog": "^1.0.0",
     "ng-infinite-scroll": "^1.3.0",
-    "normalize.css": "^5.0.0",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2",
+    "normalize.css": "^7.0.0",
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4",
     "toastr": "^2.1.2",
     "universal-analytics": "^0.4.8"
   }
